app/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# HTTP Client
requests==2.31.0

# Content Extraction
trafilatura==1.6.3
pypdf==3.17.1

# AI/LLM
openai==1.3.7

# Environment Management
python-dotenv==1.0.0

# Type Hints - UPDATED VERSION
typing-extensions==4.12.2

# Optional: Remove problematic packages or upgrade them
# aiofiles==23.2.0  # Only if needed
